+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Prevent anonymous users from accessing setup pages.
 - Firstboot, KVStore: merge old firstboot state fields.
 - tor: Use Plinth-specific instance instead of default.
+- xmpp: Split into ejabberd and jsxc modules.
 
 ## [0.12.0] - 2016-12-08
 ### Added